build: resolve 2 instances of -Wformat
exch/http/hpm_processor.cpp: In function "http_status hpm_processor_take_request(http_context*)":
exch/http/hpm_processor.cpp:265:37: warning: format "%zu" expects
argument of type "size_t", but argument 4 has type "uint64_t" {aka
"long long unsigned int"} [-Wformat=]
  265 |                                 "(%zu > %zu; http.cfg:http_rqbody_max_size)",
exch/http/mod_fastcgi.cpp: In function "http_status mod_fastcgi_take_request(http_context*)":
exch/http/mod_fastcgi.cpp:435:29: warning: format "%zu" expects
argument of type "size_t", but argument 4 has type "uint64_t" {aka
"long long unsigned int"} [-Wformat=]
  435 |                         "(%zu > %zu; http.cfg:http_rqbody_max_size)",

+++ exch/http/hpm_processor.cpp
@@ -262,8 +262,8 @@ http_status hpm_processor_take_request(http_context *phttp)
 			continue;
 		if (rq.content_len > g_rqbody_max_size) {
 			phttp->log(LV_INFO, "rejected because Content-Length too large "
-				"(%zu > %zu; http.cfg:http_rqbody_max_size)",
-				rq.content_len, g_rqbody_max_size);
+				"(%llu > %zu; http.cfg:http_rqbody_max_size)",
+				static_cast<unsigned long long>(rq.content_len), g_rqbody_max_size);
 			return http_status::bad_request;
 		}
 		if (rq.b_chunked || rq.content_len > g_rqbody_flush_size) {
